requirements.txt
fastapi==0.115.14
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.19
python-jose[cryptography]==3.4.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.2
passlib[bcrypt]==1.7.4
itsdangerous==2.1.2
Markdown==3.8.2
pydantic==2.11.7
qrcode==8.2
qrcode[pil]==8.2
aiosqlite==0.20.0
Jinja2==3.1.6
bcrypt==4.0.1